+++ assets/js/src/maps.js
@@ -126,6 +126,8 @@ var LSX_TO_Maps = {
 
 		this.resizeThis();
 
+		$map.removeClass('hidden');
+
 		$gmap = this.mapObj;
 	},